Add abstract types to separete single-target and multi-target MLJ Regressor objects.
Right now, SRRegressor and MultitargetSRRegressor subtype AbstractSRRegresor. However, much of the `src/MLJInterface.jl` code dispatches on either SRRegressor or MultitargetSRRegressor. This creates a bottleneck for defining custom MLJInterfaces and forces redefinition of much of `src/MLJInterface.jl`. The commit addresses this by introducing two new types, SingletargetAbstractSRRegressor and MultitargetAbstractSRRegressor, and changes logic in `src/MLJInterface.jl` to dispatch on these types instead.

Note: This approach is kind of janky as it solves the reuse problem using multiple inheritence (which might cause problems of its own if abused). This can be trimmed further / I'm open to rethinking this.