Generalize row collation in HallOfFame.jl.
Presently, HallOfFame is overfit to only a select few metrics that can possibly be tracked using subclasses of AbstractPopMember. This (and following) commits attempt to generalize the framework by allowing users to add arbitrary metrics in rows.

This also opens up the API to Tables.jl through the SymbolicRegressionTablesExt.jl file which will, amongst other things, allow us to save the hall-of-fame as a DataFrame.jl or a JuliaDB.